Move video/user tests to the new folder structure
