--- src/Context/Video/Module/User/Test/PhpUnit/UserModuleFunctionalTestCase.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace CodelyTv\Context\Video\Module\User\Test\PhpUnit;
-
-use CodelyTv\Test\Context\Video\PhpUnit\VideoContextFunctionalTestCase;
-
-abstract class UserModuleFunctionalTestCase extends VideoContextFunctionalTestCase
-{
-}
