+++ tests/Context/Video/Module/User/Application/Find/FindUserTest.php
@@ -2,16 +2,15 @@
 
 declare(strict_types = 1);
 
-namespace CodelyTv\Context\Video\Module\User\Tests\Behaviour;
+namespace CodelyTv\Test\Context\Video\Module\User\Application\Find;
 
 use CodelyTv\Context\Video\Module\User\Application\Find\FindUserQueryHandler;
 use CodelyTv\Context\Video\Module\User\Application\Find\UserFinder;
 use CodelyTv\Context\Video\Module\User\Domain\UserNotExist;
-use CodelyTv\Context\Video\Module\User\Test\PhpUnit\UserModuleUnitTestCase;
-use CodelyTv\Context\Video\Module\User\Test\Stub\FindUserQueryStub;
-use CodelyTv\Context\Video\Module\User\Test\Stub\UserIdStub;
-use CodelyTv\Context\Video\Module\User\Test\Stub\UserResponseStub;
-use CodelyTv\Context\Video\Module\User\Test\Stub\UserStub;
+use CodelyTv\Test\Context\Video\Module\User\UserModuleUnitTestCase;
+use CodelyTv\Test\Context\Video\Module\User\Domain\UserIdStub;
+use CodelyTv\Test\Context\Video\Module\User\Domain\UserResponseStub;
+use CodelyTv\Test\Context\Video\Module\User\Domain\UserStub;
 
 final class FindUserTest extends UserModuleUnitTestCase
 {
