+++ tests/Context/Video/Module/User/UserModuleFunctionalTestCase.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CodelyTv\Test\Context\Video\Module\User;
+
+use CodelyTv\Test\Context\Video\VideoContextFunctionalTestCase;
+
+abstract class UserModuleFunctionalTestCase extends VideoContextFunctionalTestCase
+{
+}
